requirements.txt
guardrails-ai 
ollama 
pandas 
numpy
setuptools
nltk
sentence-transformers
fastapi
"uvicorn[standard]"
jinja2